refactor(engine): move CachedStateProvider into reth-execution-cache
cached_state.rs is now just re-exports.

Amp-Thread-ID: https://ampcode.com/threads/T-019d2125-584a-700b-84e8-b4f33b96d56f
